+++ client/src/shared/Modals/index.tsx
@@ -1,0 +1,3 @@
+import ConfirmModal from "./ConfirmModal/ConfirmModal";
+
+export { ConfirmModal, FormModal };
